python/naming/requirements.txt
connexion
connexion[swagger-ui]
Flask-Caching
flask_marshmallow
flask_sqlalchemy
Flask-SQLAlchemy-Caching
folium
hupper
marshmallow-sqlalchemy
pandas
waitress
